requirements.txt
python-docx==0.8.11
openpyxl==2.5.11
xlrd==2.0.1
xlutils==2.0.0
XlsxWriter==1.4.4
pandas==1.2.4